limpieza

+++ sources/src/print/ft_print_unsint.c
@@ -35,7 +35,7 @@ static void	ft_u_itoa_hex(unsigned int nbr, t_print *tp)
 
 void	ft_print_unsint(va_list arg, t_print *tp)
 {
-	unsigned long	nbr;
+	unsigned int	nbr;
 
 	nbr = va_arg(arg, unsigned int );
 	ft_u_itoa_hex(nbr, tp);
